refactor(core): refactor/remove DatasetsApiMixin
