+++ renku/core/management/client.py
@@ -20,11 +20,10 @@
 import attr
 
 from renku.core.management.config import ConfigManagerMixin
-from renku.core.management.datasets import DatasetsApiMixin
 from renku.core.management.repository import PathMixin, RepositoryApiMixin
 from renku.core.management.storage import StorageApiMixin
 
 
 @attr.s
-class LocalClient(PathMixin, StorageApiMixin, RepositoryApiMixin, DatasetsApiMixin, ConfigManagerMixin):
+class LocalClient(PathMixin, StorageApiMixin, RepositoryApiMixin, ConfigManagerMixin):
     """A low-level client for communicating with a local Renku repository."""
