+++ renku/core/management/dataset/__init__.py
@@ -16,20 +16,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Renku management dataset logic."""
-
-from typing import Optional
-
-from renku.core import errors
-from renku.core.management.dataset.datasets_provenance import DatasetsProvenance
-from renku.core.models.dataset import Dataset
-
-
-def get_dataset(name, strict=False, immutable=False) -> Optional[Dataset]:
-    """Return a dataset based on its name."""
-    datasets_provenance = DatasetsProvenance()
-    dataset = datasets_provenance.get_by_name(name, immutable=immutable)
-
-    if not dataset and strict:
-        raise errors.DatasetNotFound(name=name)
-
-    return dataset
