+++ renku/service/controllers/datasets_create.py
@@ -16,7 +16,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Renku service datasets create controller."""
-from renku.core.commands.dataset import create_dataset
+from renku.core.commands.dataset import create_dataset_command
+from renku.core.management.dataset.request_model import ImageRequestModel
 from renku.service.cache.models.job import Job
 from renku.service.config import CACHE_UPLOADS_PATH, MESSAGE_PREFIX
 from renku.service.controllers.api.abstract import ServiceCtrl
@@ -51,8 +52,18 @@ class DatasetsCreateCtrl(ServiceCtrl, RenkuOpSyncMixin):
             set_url_for_uploaded_images(images=images, cache=self.cache, user=self.user)
         user_cache_dir = CACHE_UPLOADS_PATH / self.user.user_id
 
+        images = [
+            ImageRequestModel(
+                content_url=img["content_url"],
+                position=img["position"],
+                mirror_locally=img["mirror_locally"],
+                safe_image_paths=[user_cache_dir],
+            )
+            for img in images
+        ]
+
         return (
-            create_dataset()
+            create_dataset_command()
             .with_commit_message(self.ctx["commit_message"])
             .build()
             .execute(
@@ -63,7 +74,6 @@ class DatasetsCreateCtrl(ServiceCtrl, RenkuOpSyncMixin):
                 keywords=self.ctx.get("keywords"),
                 images=self.ctx.get("images"),
                 custom_metadata=self.ctx.get("custom_metadata"),
-                safe_image_paths=[user_cache_dir],
             )
         )
