+++ tests/core/fixtures/core_datasets.py
@@ -21,6 +21,9 @@ from contextlib import contextmanager
 
 import pytest
 
+from renku.core.management.dataset.context import DatasetContext
+from renku.core.management.dataset.usecase import add_data_to_dataset, create_dataset
+
 
 @pytest.fixture
 def dataset_responses():
@@ -59,13 +62,13 @@ def client_with_datasets(client, directory_tree, client_database_injection_manag
     person_2 = Person.from_string("P2 <p2@example.com>")
 
     with client_database_injection_manager(client):
-        client.create_dataset(name="dataset-1", keywords=["dataset", "1"], creators=[person_1])
+        create_dataset(name="dataset-1", keywords=["dataset", "1"], creators=[person_1])
 
-        with client.with_dataset(name="dataset-2", create=True, commit_database=True) as dataset:
+        with DatasetContext(name="dataset-2", create=True, commit_database=True) as dataset:
             dataset.keywords = ["dataset", "2"]
             dataset.creators = [person_1, person_2]
 
-            client.add_data_to_dataset(dataset=dataset, urls=[str(p) for p in directory_tree.glob("*")])
+            add_data_to_dataset(dataset=dataset, urls=[str(p) for p in directory_tree.glob("*")])
 
     client.repository.add(all=True)
     client.repository.commit("add files to datasets")
@@ -92,7 +95,7 @@ def get_datasets_provenance_with_injection(client_database_injection_manager):
 
     @contextmanager
     def _inner(client):
-        from renku.core.management.datasets import DatasetsProvenance
+        from renku.core.management.dataset.datasets_provenance import DatasetsProvenance
 
         with client_database_injection_manager(client):
             yield DatasetsProvenance()
